renaming, add none error correction
