+++ error_correction/none_error_correction.py
@@ -1,0 +1,19 @@
+from reedsolo import RSCodec
+
+from error_correction.generic_error_correction import GenericErrorCorrection
+
+
+class NoneErrorCorrection(GenericErrorCorrection):
+    """
+        No actual error correction
+    """
+
+    @staticmethod
+    def encode(data: bytes, redundant_bits: int) -> bytes:
+
+        return data
+
+    @staticmethod
+    def decode(data: bytes, redundant_bits: int) -> bytes:
+
+        return data
